refactor: Consolidate step implementations and update documentation
This commit consolidates all step implementations into a single file and updates
the documentation to present a unified API where users think about "steps" without
needing to understand implementation differences between different types of steps.

Key changes:

1. Code Changes:
   - Move ZFlatStep, FocusStep, and CompositeStep from specialized_steps.py to steps.py
   - Remove specialized_steps.py file completely
   - Update imports in all files to import from steps.py
   - Add necessary imports to steps.py (ImageProcessor, FocusAnalyzer)

2. Documentation Changes:
   - Remove specialized_steps.rst files from both API and concepts sections
   - Update all documentation references to use consistent terminology
   - Consolidate best practices from across the documentation into best_practices.rst
   - Update cross-references between documents
   - Present all step types as part of a unified step system
   - Ensure documentation follows a top-down approach

3. Test Changes:
   - Rename test_specialized_steps.py to test_step_factories.py
   - Update function names and docstrings in tests
   - Update pipeline names in tests

This change simplifies the codebase by having all step types in a single file
and makes the documentation more cohesive by presenting a unified API to users.
All tests are passing, confirming that the changes maintain functionality.
