+++ docs/source/concepts/step.rst
@@ -8,6 +8,12 @@ Overview
 A ``Step`` is a single processing operation that can be applied to images. It is a key component of the EZStitcher architecture.
 For an overview of the complete architecture, see :doc:`architecture_overview`.
 
+EZStitcher provides several types of steps:
+
+1. **Base Step**: The foundation for all step types, providing core functionality
+2. **Step Factories**: Pre-configured steps for common operations (ZFlatStep, FocusStep, CompositeStep)
+3. **Specialized Steps**: Steps for specific tasks (PositionGenerationStep, ImageStitchingStep)
+
 The base ``Step`` class provides:
 
 * Image loading and saving
@@ -108,7 +114,7 @@ For practical examples of how to use variable_components in different scenarios,
 
     # IMPORTANT: For Z-stack flattening, use ZFlatStep instead of raw Step with variable_components
     # This is the recommended approach for Z-stack flattening
-    from ezstitcher.core.specialized_steps import ZFlatStep
+    from ezstitcher.core.steps import ZFlatStep
 
     # Maximum intensity projection (default)
     step = ZFlatStep()  # Uses max_projection by default
@@ -118,7 +124,7 @@ For practical examples of how to use variable_components in different scenarios,
 
     # IMPORTANT: For channel compositing, use CompositeStep instead of raw Step with variable_components
     # This is the recommended approach for channel compositing
-    from ezstitcher.core.specialized_steps import CompositeStep
+    from ezstitcher.core.steps import CompositeStep
 
     # Without weights (equal weighting for all channels)
     step = CompositeStep()  # Equal weights for all channels
@@ -209,53 +215,14 @@ In this example:
 
 6. ``well_filter`` is optional and will inherit from the pipeline's context if not specified.
 
-Step Parameters Best Practices
-----------------------------
+.. _step-best-practices:
 
-When configuring step parameters, follow these best practices:
+Best Practices
+------------
 
-1. **Use Descriptive Names**:
-   - Choose clear, descriptive names for your steps
-   - This makes pipelines easier to understand and debug
+For comprehensive best practices on using steps effectively, see :ref:`best-practices-steps` in the :doc:`../user_guide/best_practices` guide.
 
-2. **Function Handling**:
-   - Use the tuple pattern ``(func, kwargs)`` for passing arguments to functions
-   - Use lists of functions for sequential processing
-   - Use dictionaries of functions with ``group_by`` for component-specific processing
-   - Use the ``stack()`` utility for adapting single-image functions
-
-3. **Variable Components**:
-   - Use ``ZFlatStep`` instead of setting ``variable_components=['z_index']`` for Z-stack flattening
-   - Use ``CompositeStep`` instead of setting ``variable_components=['channel']`` for channel compositing
-   - Leave at default ``['site']`` for most other operations
-   - Only set ``variable_components`` directly when you have a specific need not covered by specialized steps
-
-4. **Directory Management**:
-   - Always specify ``input_dir`` for the first step, using ``orchestrator.workspace_path``
-   - Let EZStitcher handle directory resolution for subsequent steps
-   - Only specify ``output_dir`` when you need a specific directory structure
-
-5. **Parameter Validation**:
-   - Ensure ``group_by`` is never the same as ``variable_components``
-   - Only use ``group_by`` with dictionary functions
-   - Verify that all required parameters are specified
-
-.. _when-to-use-specialized-steps:
-
-When to Use Specialized Steps
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-For common operations, ezstitcher provides specialized step subclasses that encapsulate the appropriate
-configuration:
-
-1. **ZFlatStep**: Use for Z-stack flattening instead of manually configuring ``variable_components=['z_index']``
-2. **FocusStep**: Use for focus detection in Z-stacks
-3. **CompositeStep**: Use for channel compositing instead of manually configuring ``variable_components=['channel']``
-
-These specialized steps provide cleaner, more readable code and ensure proper configuration. Use them with minimal parameters unless you need to override defaults.
+For information on when to use specialized steps, see :ref:`best-practices-steps` in the :doc:`../user_guide/best_practices` guide.
 
 For channel-specific processing with different functions per channel, using a raw ``Step`` with a dictionary
 of functions and ``group_by='channel'`` is the appropriate approach.
-
-Specialized steps extend the base Step class with specific functionality.
-For detailed information about specialized steps, see :doc:`specialized_steps`.
