+++ docs/source/user_guide/intermediate_usage.rst
@@ -45,8 +45,7 @@ For more control, use custom pipelines:
    from pathlib import Path
    from ezstitcher.core.pipeline_orchestrator import PipelineOrchestrator
    from ezstitcher.core.pipeline import Pipeline
-   from ezstitcher.core.steps import Step, PositionGenerationStep, ImageStitchingStep
-   from ezstitcher.core.specialized_steps import ZFlatStep, CompositeStep
+   from ezstitcher.core.steps import Step, PositionGenerationStep, ImageStitchingStep, ZFlatStep, CompositeStep
    from ezstitcher.core.image_processor import ImageProcessor as IP
 
    plate_path = Path("~/data/PlateA")  # <-- edit me
@@ -90,8 +89,7 @@ assemble the final mosaic with **best-focus** selection.
 
    from pathlib import Path
    from ezstitcher.core.pipeline import Pipeline
-   from ezstitcher.core.steps import Step, PositionGenerationStep, ImageStitchingStep
-   from ezstitcher.core.specialized_steps import ZFlatStep, FocusStep
+   from ezstitcher.core.steps import Step, PositionGenerationStep, ImageStitchingStep, ZFlatStep, FocusStep
    from ezstitcher.core.image_processor import ImageProcessor as IP
 
    # --- reusable position pipeline ---------------------------------
